refactor(app2): move evm balances
